document service & fix bog

+++ database/models/document.model.go
@@ -12,7 +12,7 @@ type Document struct {
 	UserID          string    `gorm:"type:varchar(40);not null"`
 	Title           string    `gorm:"type:varchar(255);not null" `
 	Category        string    `gorm:"type:varchar(100);not null" `
-	Ratings         float64   `gorm:"type:decimal(3,2);default:0" `
+	Ratings         int       `gorm:"type:decimal(3,2);default:0" `
 	Description     string    `gorm:"type:text" `
 	Status          string    `gorm:"type:varchar(50);not null" `
 	PublicationDate time.Time `gorm:"not null"`
